Use common result.json parse func on Summary page
We remove the parseJSONData() function in the Summary page, and use the
queryIterations() and parseIterationData() common methods used by other
pages. To make this work, we first need to ensure the controller field
is returned with result data in models/fetchResults().

We also added support for gracefully handling missing result.json files
which return a 404 by logging to the console the 404 error, and then
checking for all the expected data to be present in the render() method
of the Summary page.