Update README for using dashboard playbooks
We add the missing `graphql` configuration option to the `config.js.j2`
template.

We also add a convenience link to the `config.json.j2` to make it easier
to use the scripts, and add an ignore the generated `dist` directories.

Further, we now ignore patch files from git format-patch, and ignore
`run-unittests.log` files.