Dashboard support for endpoint configuration API
For everyone who's been wondering where I was heading with pbench #2024
("Add endpoint configuration query"), with dashboard #95 merged ("Dashboard
to work with Server APIs" merged I'm ready to push this into a draft PR to
get some visibility.

This replaces the loading of a static file from the template .html file
and automatically configures the dashboard to match the server web
endpoint and index configuration. Although the form of the <script> tag
is different, in terms of effective Javascript logic and HTTP traffic it's
effectively identical ... the GET is just pulling the body of the endpoints
dictionary from the Pbench server rather than from Apache.