Third iteration: create `endpoints.ts`
The result of looking at UMI documentation, after finding out that
"public" is apparently hardwired, along with "mock", and seeing a
reference to automatic use of ".ts" files for mocking, I tried replacing
my static mocked endpoints.mock.json file with a GET mock in a
mock/endpoints.ts file. This eliminates the need to ejs-template
the fetch in documents.ejs, and it works for start:mock, but the
e2e tests are still mysteriously failing, and the output isn't helping
me to figure out what's happening.