A few more fixes
With this, the start:mock environment can get past the controllers page.

Gurbir is working on the e2e mocking problems, and we'll merge at
some point when he's got the magic stable.

Note that it doesn't appear that the timeseries data query had been
mocked previously, and I added a placeholder to make this explicit.
Apparently the e2e tests don't cover the Compare feature so the lack
didn't cause a failure. (I noticed it in poking around in the start:mock
environment to see if I could cover all the use cases.)