Pass logged-in user to controller query
Pass the logged-in user state through from the controller page to the
service and on to the back end.

This appears to work on my server VM, although no controllers will be
reported as the datasets are currently all owned by "pbench" and with
the dashboard's insistence on valid email addresses there's no way to
create a user that the server will see as "pbench". Cie la vie. For the
jam session we'll upload datasets owned by actual users.

Also, note that my e2e tests aren't working. Ironically, the controllers
suite, the only one that should be affected by my changes, is the only
one that's passing. I'm hoping this is something messed up in my local
branch and they'll pass on GitHub... if not, I'll investigate further.